+++ Cargo.toml
@@ -1,2 +1,2 @@
 [workspace]
-members = ["gonk", "gonk-player", "gonk-database", "gonk-test-database"]
+members = ["gonk", "gonk-player","gonk-database"]
