+++ gonk-database/Cargo.toml
@@ -3,15 +3,8 @@ name = "gonk-database"
 version = "0.1.0"
 edition = "2021"
 
-authors = ["Bay"]
-description = "Database for gonk"
-repository = "https://github.com/zX3no/gonk"
-readme = "README.md"
-license = "MIT"
-
-
 [dependencies]
-walkdir = "2"
+memmap2 = "0.5.5"
 rayon = "1.5.3"
-rusqlite = { version = "0.28.0", features = ["bundled"] }
-gonk-player = {path = "../gonk-player"}
+walkdir = "2.3.2"
+symphonia = { version = "0.5.0", features = ["mp3", "isomp4", "alac", "aac"] }
