+++ gonk-database/src/main.rs
@@ -1,0 +1,9 @@
+use gonk_database::*;
+
+fn main() {
+    init();
+
+    bench(|| {
+        let _item = ids(&[1000, 232, 314, 414, 11, 2001]);
+    });
+}
