+++ gonk-player/Cargo.toml
@@ -12,6 +12,7 @@ license = "MIT"
 [dependencies]
 cpal = {git = "https://github.com/zX3no/cpal"}
 symphonia = { version = "0.5.0", features = ["mp3", "isomp4", "alac", "aac"] }
+gonk-database = {path = "../gonk-database"}
 
 [target.'cfg(unix)'.dependencies]
 gag = "1.0.0"
