--- gonk-test-database/Cargo.toml
@@ -1,10 +1,0 @@
-[package]
-name = "gonk-test-database"
-version = "0.1.0"
-edition = "2021"
-
-[dependencies]
-memmap2 = "0.5.5"
-rayon = "1.5.3"
-walkdir = "2.3.2"
-symphonia = { version = "0.5.0", features = ["mp3", "isomp4", "alac", "aac"] }
