+++ gonk/src/settings.rs
@@ -1,5 +1,4 @@
 use crate::{log, widgets::*, Frame, Input};
-use gonk_database::query;
 use gonk_player::{Device, DeviceTrait, Index, Player};
 use tui::{
     layout::Rect,
@@ -14,26 +13,28 @@ pub struct Settings {
 
 impl Settings {
     pub fn new() -> Self {
-        let default_device = gonk_player::default_device();
-        let wanted_device = query::playback_device();
+        // let default_device = gonk_player::default_device();
+        // let wanted_device = gonk_database::playback_device();
 
-        let devices = gonk_player::audio_devices();
+        // let devices = gonk_player::audio_devices();
 
-        let current_device = if devices
-            .iter()
-            .flat_map(DeviceTrait::name)
-            .any(|x| x == wanted_device)
-        {
-            wanted_device
-        } else {
-            let name = default_device.name().unwrap();
-            query::set_playback_device(&name);
-            name
-        };
+        // let current_device = if devices
+        //     .iter()
+        //     .flat_map(DeviceTrait::name)
+        //     .any(|x| x == wanted_device)
+        // {
+        //     wanted_device
+        // } else {
+        //     let name = default_device.name().unwrap();
+        //     gonk_database::set_playback_device(&name);
+        //     name
+        // };
 
         Self {
-            devices: Index::new(devices, Some(0)),
-            current_device,
+            // devices: Index::new(devices, Some(0)),
+            // current_device,
+            devices: Index::default(),
+            current_device: String::new(),
         }
     }
 }
